refactor(proxy): split proxy_wiring monolith into focused submodules
Mechanical move, no logic change (backlog #7 step 3). The module
root shrinks from 5813 to 2482 LOC and keeps the LoricaProxy struct,
its spawn helpers, the cache statics, and the ProxyHttp trait impl;
the rest moves to:

- config.rs: ProxyConfig/ProxyConfigGlobals/RouteEntry/SmoothWrr +
  route-table construction + find_route + compute_pool_size
- lb.rs: EwmaTracker + CircuitBreaker + BackendConnections re-export
- context.rs: RequestCtx
- filters.rs: the 22 request_filter stage methods +
  write_error_response + persist_waf_event + WAF body-cap recorder
- worker_rpc.rs: spawn_worker_rpc_listener + WorkerMetricsCtx + the
  config-reload prepare/commit/abort and metrics RPC handlers

Bodies verbatim; only visibilities (pub(super)/pub(crate) where
cross-module), imports, and root re-exports changed. The public
lorica::proxy_wiring::* surface is preserved via pub use (every
external reference grepped and kept valid); tests.rs and
cert_reload_commit_tests.rs are byte-identical and compile through
cfg(test) re-exports.
